Changes made in polygon subgraphs
